requirements.txt
autopep8==2.0.0
mysql-connector-python==8.0.31
protobuf==3.20.1
pycodestyle==2.9.1
tomli==2.0.1
